backend/requirements.txt
fastapi
uvicorn
qdrant-client
sentence-transformers
pydantic
python-dotenv
requests
